+++ Tests/Functional/Core/CoreSpecial/BatchSimulation.cpp
@@ -22,7 +22,7 @@ bool TestBatchSimulation()
 
     const int n_batches = 9;
     const double threshold = 2e-10;
-    for(size_t i_batch=0; i_batch<n_batches; ++i_batch) {
+    for(int i_batch=0; i_batch<n_batches; ++i_batch) {
         const std::unique_ptr<GISASSimulation> batch(simulation->clone());
         ThreadInfo threadInfo;
         threadInfo.n_threads = 1;
